Move filesize into SummaryFormatter
